# Fix function-no-unknown performance