--- a/.changeset/shy-tools-rest.md
+++ b/.changeset/shy-tools-rest.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+- Fixed: `function-no-unknown` performance