--- a/.changeset/shy-tools-rest.md
+++ b/.changeset/shy-tools-rest.md
@@ -2,4 +2,4 @@
 "stylelint": patch
 ---
 
-- Fixed: `function-no-unknown` performance
+Fixed: `function-no-unknown` performance