--- a/lib/rules/function-no-unknown/index.js
+++ b/lib/rules/function-no-unknown/index.js
@@ -62,6 +62,8 @@ const rule = (primary, secondaryOptions) => {
 		root.walkDecls((decl) => {
 			const { value } = decl;
 
+			if (!value.includes('(')) return;
+
 			if (!isStandardSyntaxValue(value)) return;
 
 			/**