<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

Ref #6869 

> Is there anything in the PR that needs further explanation?

Before:
Mean: 104.89167158823528 ms
Deviation: 9.601897806542654 ms

After:
Mean: 71.84089116666665 ms
Deviation: 10.215987936599932 ms