@jeddy3 2023-06-30
> LGTM! How about adding a changelog entry like this?

Oops, I'm rusty 😆 . I forgot to do that.